+++ templates/base/__sidenav.html.twig
@@ -85,7 +85,7 @@
                         <a href="{{ path('team_abteilungen') }}">Abteilungen</a>
                     </li>
                     <li>
-                        <a href="{{ path('team_mitglieder') }}">Datenschutzkoordinatoren</a>
+                        <a href="{{ path('team_mitglieder') }}">Benutzer</a>
                     </li>
                     <li>
                         <a href="{{ path('akademie_admin') }}">Akademie</a>
