+++ templates/team/academy.html.twig
@@ -3,8 +3,9 @@
 {% block title %}{{ title }}{% endblock %}
 {% block body %}
 
-    <div class="container">
 
+    <div class="container">
+        <h2 class="h2-responsive">{{ title }}</h2>
         <ul class="nav nav-tabs" id="myTab" role="tablist">
             <li class="nav-item">
                 <a class="nav-link active" id="home-tab" data-toggle="tab" href="#courses" role="tab"
@@ -28,9 +29,9 @@
                        class="btn btn-primary btn-sm loadContent">Mitglied hinzufügen</a>
                     <ul class="list-group">
                         {% for d in data %}
-                            <li class="list-group-item">{{ d.email }} <a class="btn btn-sm btn-danger float-md-right"
-                                                                         onClick="return confirm('{% trans %}Möchten Sie diese Aktion wirklich durchführen?{% endtrans %}')"
-                                                                         href="{{ path('akademie_mitglieder_remove',{'id':d.id}) }}">Entfernen</a>
+                            <li class="list-group-item">{{ d.email }}<a class="btn btn-sm btn-danger float-md-right"
+                                                                        onClick="return confirm('{% trans %}Möchten Sie diese Aktion wirklich durchführen?{% endtrans %}')"
+                                                                        href="{{ path('team_mitglieder_remove',{'id':d.id, 'type':'academy'}) }}">Entfernen</a>
                             </li>
                         {% endfor %}
                     </ul>
@@ -42,5 +43,15 @@
 
 {% endblock %}
 
-{% block javascript %}
+{% block javascripts %}
+    <script>
+        var url = document.location.toString();
+        if (url.match('#')) {
+            $('.nav-tabs a[href="#' + url.split('#')[1] + '"]').tab('show');
+        }
+
+        $('.nav-tabs a').on('shown.bs.tab', function (e) {
+            window.location.hash = e.target.hash;
+        });
+    </script>
 {% endblock %}
