+++ templates/team/member.html.twig
@@ -6,7 +6,9 @@
     <div class="container">
         <h2 class="h2-responsive">{{ title }}</h2>
         <a href="{{ path('team_mitglieder_create', {'title':'Mitglied hinzufügen', 'type':'odc'}) }}"
-           class="btn btn-primary btn-sm loadContent">Mitglied hinzufügen</a>
+           class="btn btn-primary btn-sm loadContent">Benutzer hinzufügen</a>
+        <a href="{{ path('akademie_admin') }}#user"
+           class="btn btn-cyan btn-sm">Akademiebenutzer</a>
         <div class="card card-body">
 
             <ul class="list-group">
@@ -18,13 +20,13 @@
                                 <span class="badge-success p-1 rounded">Admin</span>{% endif %} {% if d.akademieUser == data %}
                                 <span class="badge-warning p-1 rounded">Akademie</span>{% endif %} {% if d.team == data %}
                                 <span class="badge-info p-1 rounded">Koordinator</span>{% endif %}
-                            </div>
-                            <a class="btn btn-sm btn-info float-md-right"
-                               href="{{ path('team_mitglieder_admin', {'id':d.id}) }}">{% if d.adminUser %}{% trans %}Admin entfernen{% endtrans %}{% else %}{% trans %}Admin ernennen{% endtrans %}{% endif %}</a>
-                            <a class="btn btn-sm btn-danger float-md-right"
-                               href="{{ path('team_mitglieder_remove',{'id':d.id}) }}">{% trans %}Entfernen{% endtrans %}</a>
-                        </li>
-                    {% endfor %}
+                        </div>
+                        <a class="btn btn-sm btn-info float-md-right"
+                           href="{{ path('team_mitglieder_admin', {'id':d.id}) }}">{% if d.adminUser %}{% trans %}Admin entfernen{% endtrans %}{% else %}{% trans %}Admin ernennen{% endtrans %}{% endif %}</a>
+                        <a class="btn btn-sm btn-danger float-md-right"
+                           href="{{ path('team_mitglieder_remove',{'id':d.id,'type':'odc'}) }}">{% trans %}Entfernen{% endtrans %}</a>
+                    </li>
+                {% endfor %}
                 </ul>
 
                 <h5 class="h5-responsive mt-3">{% trans %}Offene Einladungen{% endtrans %}</h5>
